+++ src/main/java/com/panacademy/squad7/bluebank/services/AddressesService.java
@@ -1,20 +1,18 @@
 package com.panacademy.squad7.bluebank.services;
 
-import com.panacademy.squad7.bluebank.domain.models.Address;
-import org.springframework.stereotype.Service;
-
 import java.util.List;
-import java.util.Optional;
+
+import com.panacademy.squad7.bluebank.domain.models.Address;
 
 public interface AddressesService {
 
-    Address  create(Address address);
+    Address create(Address address);
 
-    Address update(Address address, Integer id);
+    Address update(Address address, Long id);
 
-    void  delete(Integer id);
+    void delete(Long id);
 
-    Address findById(Integer id);
+    Address findById(Long id);
 
     List<Address> findAll();
 }
